Refactor DataPool GenServer State
I don't want to work with a naked pid anymore with the data pool
because it doesn't fit well when trying to implment protocols as
well as supporting more advanced features.  This refactor pushes
the state into it's own structure and the DataPool struct now
manages it's own pid.  If you were using this before and weren't
trying to send any goofy messages to the pid this shouldn't break
anything.