Adding Pool State Status
Changing the contract with how push and pop with with the new
concept of the pool status.  After analyzing how I want to use
the pool as a buffer it has become clear that it is beneficial
for it to track a status that matches how the collectable
protocol works.