Fix Race Condition Bug
This was an elusive to track down bug; however, it would show itself
as a "deadlock" in which the stream buffer would never finish for
long running tasks.  What ended up being the problem is that when
the stream was updated to `done` or `halt` it was sending the incorrect
value to `GenServer.reply`.

`GenServer.reply` expects a two pair tuple of {pid,ref}, however what
it was being given was a two pair tuple of {{pid, ref}, any}.  The
GenServer library happily takes this two pair tuple, attempts to use it
and swallows the resulting error from trying to use incorrect data.

I have written a test which properly exercises this "deadlock" scenario
and have updated the code to use the correct part of the tuple to send
replies to the blocked producers that the state has changed to `done` or
`halt`.